+++ BTL-Java-Group2/src/Model/Phong.java
@@ -3,7 +3,7 @@ package Model;
 import java.io.File;
 
 public class Phong {
-    private int id;
+    private Long id;
     private DiaChi diaChi;
     private double gia;
     private String moTa;
@@ -15,7 +15,7 @@ public class Phong {
     public Phong() {
     }
 
-    public Phong(int id, DiaChi diaChi, double gia, String moTa, double dienTich, File hinhAnh, ChuPhong chu, KhachHang khach) {
+    public Phong(Long id, DiaChi diaChi, double gia, String moTa, double dienTich, File hinhAnh, ChuPhong chu, KhachHang khach) {
         this.id = id;
         this.diaChi = diaChi;
         this.gia = gia;
@@ -26,11 +26,11 @@ public class Phong {
         this.khach = khach;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
